// packages/backend/src/MoodleNetDomain/services/UserAuth/UserAuth.graphql.gen.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: Id;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Never: never;
  Empty: {};
  DateTime: Date;
};
// ...
export type SimpleResponse = {
  __typename: 'SimpleResponse';
  success: Scalars['Boolean'];
  message?: Maybe<Scalars['String']>;
};

export type CreateSession = {
  __typename: 'CreateSession';
  jwt?: Maybe<Scalars['String']>;
  message?: Maybe<Scalars['String']>;
};

export type UserSession = {
  __typename: 'UserSession';
  username: Scalars['String'];
  email: Scalars['String'];
  changeEmailRequest?: Maybe<Scalars['String']>;
  userId: Scalars['String'];
  profileId?: Maybe<Scalars['ID']>;
};